services/region-detector/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
minio>=7.2.5
joblib>=1.3.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.0.0
pytest>=8.0.0
